Add files via upload
This is a zip file for zombie shooter game. this file includes all the resources and classes etc.
The project has database connected to it.
It is has additional features which i added to it.
Like players to choose from,
store highscore and compare for new highscore.
you can switch guns aswell.
there is a count of lives for player.